Only create db for writers

+++ benchmarks/vote/clients/hybrid.rs
@@ -15,7 +15,7 @@ pub struct W<'a> {
     mc: memcached::Client,
 }
 
-pub fn setup(memcached_dbn: &str, mysql_dbn: &str) -> Pool {
+pub fn setup(memcached_dbn: &str, mysql_dbn: &str, write: bool) -> Pool {
     use mysql::Opts;
 
     let mc = memcached::Client::connect(&[(&format!("tcp://{}", memcached_dbn), 1)],
@@ -23,34 +23,33 @@ pub fn setup(memcached_dbn: &str, mysql_dbn: &str) -> Pool {
         .unwrap();
 
     let addr = format!("mysql://{}", mysql_dbn);
-    // we need to do this dance to avoid using the DB early (which will crash us if it doesn't
-    // exist)
-    let db = &addr[addr.rfind("/").unwrap() + 1..];
-    let opts = Opts::from_url(&addr[0..addr.rfind("/").unwrap()]).unwrap();
+    if write {
+        // clear the db (note that we strip of /db so we get default)
+        let db = &addr[addr.rfind("/").unwrap() + 1..];
+        let opts = Opts::from_url(&addr[0..addr.rfind("/").unwrap()]).unwrap();
+        let pool = mysql::Pool::new(opts).unwrap();
+        let mut conn = pool.get_conn().unwrap();
+        if conn.query(format!("USE {}", db)).is_ok() {
+            conn.query(format!("DROP DATABASE {}", &db).as_str()).unwrap();
+        }
+        conn.query(format!("CREATE DATABASE {}", &db).as_str()).unwrap();
 
-    // Check whether database already exists, or whether we need to create it
-    let pool = mysql::Pool::new(opts).unwrap();
-    let mut conn = pool.get_conn().unwrap();
-    if conn.query(format!("USE {}", db)).is_ok() {
-        conn.query(format!("DROP DATABASE {}", &db).as_str()).unwrap();
+        // allow larger in-memory tables (4 GB)
+        pool.prep_exec("SET max_heap_table_size = 4294967296", ()).unwrap();
+
+        // create tables with indices
+        pool.prep_exec("CREATE TABLE art (id bigint, title varchar(255), votes bigint, \
+                        PRIMARY KEY USING HASH (id)) ENGINE = MEMORY",
+                       ())
+            .unwrap();
+        pool.prep_exec("CREATE TABLE vt (u bigint, id bigint, PRIMARY KEY USING HASH (u, id), \
+                        KEY id (id)) ENGINE = MEMORY",
+                       ())
+            .unwrap();
     }
-    conn.query(format!("CREATE DATABASE {}", &db).as_str()).unwrap();
-
-    // allow larger in-memory tables (4 GB)
-    pool.prep_exec("SET max_heap_table_size = 4294967296", ()).unwrap();
-
-    // create tables with indices
-    pool.prep_exec("CREATE TABLE art (id bigint, title varchar(255), votes bigint, \
-                    PRIMARY KEY USING HASH (id)) ENGINE = MEMORY",
-                   ())
-        .unwrap();
-    pool.prep_exec("CREATE TABLE vt (u bigint, id bigint, PRIMARY KEY USING HASH (u, id), \
-                    KEY id (id)) ENGINE = MEMORY",
-                   ())
-        .unwrap();
 
     Pool {
-        sql: pool,
+        sql: mysql::Pool::new(Opts::from_url(&addr).unwrap()).unwrap(),
         mc: Some(mc),
     }
 }
